# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Example API",
        "version": "1.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "content": {
                        "applocation/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "User email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully"
                    }
                }
            }
        },
        "api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User logout",
                "operationId": "86501f82c5fa28faf0bb63532f8cd163",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {}
                ]
            }
        },
        "api/resetPassword": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Reset password",
                "operationId": "c4f8dd3d5efd059744e21ffe681666db",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "user email",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reset password"
                    }
                }
            }
        },
        "/api/groups": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "Get groups list",
                "operationId": "ac30b3897794f3ee24af2af44d36ce5d",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "groups": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "Group id",
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "description": "Group name",
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "description": "Creating date",
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "updated_at": {
                                                        "description": "Updating date",
                                                        "type": "string",
                                                        "format": "date"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Groups"
                ],
                "summary": "Store group",
                "operationId": "7e6685b499160ee7d51b1389568e86de",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Group name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/groups/{group}": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "Show group",
                "operationId": "e2d6a534a8d1765ee2f8b583628fd816",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "description": "Group id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "group": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "Group id",
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "description": "Group name",
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "description": "Creating date",
                                                        "type": "string",
                                                        "format": "date"
                                                    },
                                                    "updated_at": {
                                                        "description": "Updating date",
                                                        "type": "string",
                                                        "format": "date"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Groups"
                ],
                "summary": "Update group",
                "operationId": "e92f78d7006c10c6ee089b9bb58153e6",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "description": "Group id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Group name",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Groups"
                ],
                "summary": "Delete group",
                "operationId": "b94c1c749a1c9c46a0121a71adbde115",
                "parameters": [
                    {
                        "name": "group",
                        "in": "path",
                        "description": "Group id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    }
}